fix cpu omit _szpflag8 (T.Yui)
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/trunk@310 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ i286c/i286c_f6.c
@@ -149,7 +149,7 @@ I286_F6 _idiv_ea8(UINT op) {
 	if (src) {
 		r = tmp / src;
 		if (!((r + 0x80) & 0xff00)) {
-			I286_AL = r;
+			I286_AL = (UINT8)r;
 			I286_AH = tmp % src;
 			return;
 		}
